# docs/swagger.yaml
basePath: /api
definitions:
  model.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.Comment:
    properties:
      author_id:
        type: integer
      author_name:
        description: 冗余字段，减少查询
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      parent_id:
        description: 为空表示一级评论，否则为回复
        type: integer
      post_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  model.CreateCommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
      post_id:
        type: integer
    required:
    - content
    - post_id
    type: object
...
  model.ListResponse:
    properties:
      data: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.LoginResponse:
    properties:
      expire_at:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  model.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  model.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
    type: object
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
  utils.Response:
    properties:
      code:
        $ref: '#/definitions/utils.ResponseCode'
      data: {}
      message:
        type: string
    type: object
  utils.ResponseCode:
    enum:
    - 0
    - 400
    - 401
    - 403
    - 404
    - 500
    - 1001
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeInternalError
    - CodeBusinessError
host: localhost:8080
info:
  contact:
    name: 博客作者
    url: http://localhost:8080
  description: 基于 Gin + GORM + MySQL + Redis 的个人博客后端 RESTful API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080
  title: 个人博客 API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录凭证
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "1001":
          description: ""
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 认证
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户资料
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新当前登录用户资料
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "1001":
          description: ""
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 用户注册
      tags:
      - 认证
  /auth/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    badge_count:
                      type: integer
                    category_count:
                      type: integer
                    comment_count:
                      type: integer
                    post_count:
                      type: integer
                    tag_count:
                      type: integer
                    user_count:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 管理员获取站点统计
      tags:
      - 认证
  /auth/users:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 管理员获取用户列表
      tags:
      - 认证
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Category'
                  type: array
              type: object
      summary: 获取分类列表
      tags:
      - 分类
    post:
      consumes:
      - application/json
      parameters:
      - description: 分类名称
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建分类（管理员）
      tags:
      - 分类
  /categories/{id}:
    delete:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除分类（管理员）
      tags:
      - 分类
    put:
      consumes:
      - application/json
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类名称
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新分类（管理员）
      tags:
      - 分类
  /comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论
  /comments/{id}:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论
...
  /posts/{id}/comments:
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Comment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取文章评论列表
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    description: 请输入 "Bearer {token}"，例如：Bearer eyJhbGciOiJIUzI1NiIs...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"